requirements.txt
requests
beautifulsoup4
langchain
langchain-community
langchain-openai
python-dotenv
pypdf2
streamlit>=1.28.0
matplotlib>=3.5.0  
plotly>=5.0.0      